requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
pyee==7.0.4
slackclient==2.9.3
slackeventsapi==2.2.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.3
